LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-items-baseline-column-vert-lr-table-item-expected.txt

two
lines
hello

PASS #target > * 1
FAIL #target > * 2 assert_equals:
<div data-offset-x="59">hello</div>
offsetLeft expected 59 but got 60
